reduce duplication and cleanup
