+++ tests/test_interaction.py
@@ -17,6 +17,8 @@ from parcels.interaction.neighborsearch import (
 )
 from parcels.interaction.neighborsearch.basehash import BaseHashNeighborSearch
 from parcels.particle import ScipyInteractionParticle, ScipyParticle, Variable
+from tests.common_kernels import DoNothing
+from tests.utils import create_fieldset_unit_mesh
 
 ptype = {"scipy": ScipyInteractionParticle, "jit": JITParticle}
 
@@ -36,20 +38,6 @@ def DummyMoveNeighbor(particle, fieldset, time, neighbors, mutator):
     mutator[neighbor_id].append((f, ()))
 
     pass
-
-
-def DoNothing(particle, fieldset, time):
-    pass
-
-
-def create_fieldset_unit_mesh(xdim=20, ydim=20, mesh="spherical"):
-    """Standard unit mesh fieldset."""
-    lon = np.linspace(0.0, 1.0, xdim, dtype=np.float32)
-    lat = np.linspace(0.0, 1.0, ydim, dtype=np.float32)
-    U, V = np.meshgrid(lat, lon)
-    data = {"U": np.array(U, dtype=np.float32), "V": np.array(V, dtype=np.float32)}
-    dimensions = {"lat": lat, "lon": lon}
-    return FieldSet.from_data(data, dimensions, mesh=mesh)
 
 
 @pytest.fixture
